requirements.txt
Pillow>=10.2.0
pillow-heif>=0.13.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
flask>=3.0.0
cryptography>=41.0.0
werkzeug>=3.0.0
